[reduce.c] Clean up 'ininterval'
This function has two for-loops with only a semicolon as the body. This
is mystifying, because the only side-effect I can think to be relevant
is the increment of 'first'/'last'. Maybe that's it, and would probably
be better phrased as something like:

```c
first = 1;
while ((first < d) && !done[grav[first]]) {
    first++;
}
```

and

```c
last = first;
while ((last < d) && done[grav[last + 1]]) {
    last++;
}
```