[build.sh] Add function to build with UBSan
UndefinedBehaviorSanitizer is a tool for clang that detects when
undefined behavior (UB) happens in a C/C++ program. This is very useful,
because UB is prevalent in C and leads to subtle bugs.

In particular, this works out nicely because it detects an out-of-bound
array access that was crashing RSST/anc/ring16.conf! Now, all I have to
do is figure out how to avoid that problem.

I think I could possibly fix this by increasing the size of the
parameter, but I need to investigate more to see if that will work.
(Also, I need to figure out how much larger the array should be)